Major Source Update
So major updates!

I decided to try Kotlin, mainly for null safety and its many other convenient and interesting features.

I've fixed the SoundCloud integration to a point. With the testing I've done, I haven't seen any issues as of yet, but please put in an issue request if anything breaks and I'll look into it. With the SoundCloud integration back, I've added a song queue to play both SoundCloud tracks and local songs on a computer (check snippets for visuals). Playlists are supported now in addition to adding single songs from SoundCloud (note: when adding, if the song is not streamable, its link is broken, or an exception is thrown, that track will be excluded from the song list generated).

Fixed some bugs with the ControlP5 GUI updating values upon event triggers as well as keypresses.

*Note: The only issue I've encountered with the song queue is playing songs in sequence or selecting a song at an index and playing it. For some reason, advancing to the next track or playing the selected song sometimes skips one or more songs (notably more frequent when playing SoundCloud songs, especially if it's a playlist. I will be working on a fix for this, and hopefully it I'll be able to patch it soon!

I hope you enjoy! Listen on!